backend/requirements.txt
fastapi
uvicorn
faster_whisper
python-multipart
python-dotenv
